fatlabel: Optimize size of the binary
Move implementation of the alloc_rootdir_entry() function from the check.c
to boot.c and do not link unneeded objects files into fatlabel binary.

This would decrease size of the fatlabel binary nearly by half. With -g -O2
size before was about 194131 bytes and after just 104497 bytes.
