+++ src/Makefile.am
@@ -20,11 +20,12 @@ AM_CFLAGS = -Wall -Wextra -Wno-sign-compare -Wno-missing-field-initializers \
 sbin_PROGRAMS = fsck.fat mkfs.fat fatlabel
 noinst_PROGRAMS = testdevinfo
 
-fscklabel_common_sources = boot.c boot.h check.c check.h common.c common.h \
-			   fat.c fat.h file.c file.h io.c io.h lfn.c lfn.h \
-			   charconv.c charconv.h msdos_fs.h \
+fscklabel_common_sources = boot.c boot.h common.c common.h \
+			   fat.c fat.h io.c io.h msdos_fs.h \
 			   fsck.fat.h endian_compat.h
-fsck_fat_SOURCES = fsck.fat.c $(fscklabel_common_sources)
+fsck_fat_SOURCES = check.c check.h file.c file.h fsck.fat.c \
+		   charconv.c charconv.h lfn.c lfn.h \
+		   $(fscklabel_common_sources)
 fatlabel_SOURCES = fatlabel.c $(fscklabel_common_sources)
 
 mkfs_common_sources = device_info.c device_info.h \
