+++ src/boot.h
@@ -34,4 +34,13 @@ off_t find_volume_de(DOS_FS * fs, DIR_ENT * de);
 
 /* Reads the boot sector from the currently open device and initializes *FS */
 
+
+off_t alloc_rootdir_entry(DOS_FS * fs, DIR_ENT * de, const char *pattern, int gen_name);
+
+/* Allocate a free slot in the root directory for a new file. If gen_name is
+   true, the file name is constructed after 'pattern', which must include a %d
+   type format for printf and expand to exactly 11 characters. The name
+   actually used is written into the 'de' structure, the rest of *de is cleared.
+   The offset returned is to where in the filesystem the entry belongs. */
+
 #endif
