+++ src/check.h
@@ -23,14 +23,6 @@
 #ifndef _CHECK_H
 #define _CHECK_H
 
-off_t alloc_rootdir_entry(DOS_FS * fs, DIR_ENT * de, const char *pattern, int gen_name);
-
-/* Allocate a free slot in the root directory for a new file. If gen_name is
-   true, the file name is constructed after 'pattern', which must include a %d
-   type format for printf and expand to exactly 11 characters. The name
-   actually used is written into the 'de' structure, the rest of *de is cleared.
-   The offset returned is to where in the filesystem the entry belongs. */
-
 int scan_root(DOS_FS * fs);
 
 /* Scans the root directory and recurses into all subdirectories. See check.c
